requirements.txt
pillow
pdf2image
pytesseract
numpy
opencv-python